+++ framework/domain/src/main/java/is/codion/framework/domain/entity/attribute/ValueAttributeDefinition.java
@@ -102,6 +102,17 @@ public sealed interface ValueAttributeDefinition<T> extends AttributeDefinition<
 	List<Item<T>> items();
 
 	/**
+	 * @return true if a default value has been set for this attribute
+	 */
+	boolean hasDefaultValue();
+
+	/**
+	 * @return the default value for this attribute, if no default value has been set null is returned
+	 * @see #hasDefaultValue()
+	 */
+	@Nullable T defaultValue();
+
+	/**
 	 * Validates the value of this attribute as found in the given entity.
 	 * <p>Note: When validating non-nullable attributes during entity insertion
 	 * (when the entity does not exist), null values are allowed for:
@@ -183,6 +194,20 @@ public sealed interface ValueAttributeDefinition<T> extends AttributeDefinition<
 		 * @throws IllegalArgumentException in case the valid item list contains duplicate values
 		 */
 		B items(List<Item<T>> items);
+
+		/**
+		 * Sets the default value for this attribute, overrides the underlying column default value, if any
+		 * @param defaultValue the value to use as default
+		 * @return this builder instance
+		 */
+		B defaultValue(T defaultValue);
+
+		/**
+		 * Sets the default value supplier, use in case of dynamic default values.
+		 * @param supplier the default value supplier
+		 * @return this builder instance
+		 */
+		B defaultValue(ValueSupplier<T> supplier);
 
 		@Override
 		ValueAttributeDefinition<T> build();