+++ changelog.md
@@ -35,6 +35,7 @@ Codion Change Log
 - DomainModel composition improved, Entities.Configurable.add(entities, entityTypes) added, dependencies now automatically included.
 - ValueAttributeDefinition, formatting related things moved back to AttributeDefinition.
 - AttributeDefinition.nullable() moved to ValueAttributeDefinition and ForeignKeyDefinition.
+- AttributeDefinition.defaultValue() moved to ValueAttributeDefinition.
 ### is.codion.framework.model
 - EntityTableModel.EntityTableEditor added along with AbstractEntityTableModel.AbstractAbstractEntityTableEditor.
 - AbstractEntityTableModel.replaceEntities() removed, now relies on FilterModel.Items.replace() instead.